The solutions to the exercise for day 5.
Most probably quite flawed.
The net/http and open-uri versions should (and do) return a higher result, as I have made no attempt to filter out the HTML.
The problem comes with the Hpricot and Nokogiri versions. Both should return the same value, as I ask for the <body> content (with .as_plain_text and .content respectively); in any case, Nokogiri should return a higher number if .content returns HTML tags (and some contain it, say in title attributes). But it is Hpricot who yields a higher number than Nokogiri.